Rename sthGetter to sthgetter to match package name
